[test]: add test for project overview components
