+++ packages/sqle/src/page/ProjectOverview/component/ProjectScore/index.test.tsx
@@ -1,0 +1,138 @@
+import { superRender } from '../../../../testUtils/customRender';
+import ProjectScore from '.';
+import projectOverview from '../../../../testUtils/mockApi/projectOverview';
+import { act, cleanup, fireEvent, screen } from '@testing-library/react';
+import { mockUseCurrentProject } from '@actiontech/shared/lib/testUtil/mockHook/mockUseCurrentProject';
+import { mockProjectInfo } from '@actiontech/shared/lib/testUtil/mockHook/data';
+import { mockUseCurrentUser } from '@actiontech/shared/lib/testUtil/mockHook/mockUseCurrentUser';
+import EventEmitter from '../../../../utils/EventEmitter';
+import EmitterKey from '../../../../data/EmitterKey';
+import { useNavigate } from 'react-router-dom';
+import { createSpySuccessResponse } from '@actiontech/shared/lib/testUtil/mockApi';
+import eventEmitter from '../../../../utils/EventEmitter';
+
+jest.mock('react-router-dom', () => {
+  return {
+    ...jest.requireActual('react-router-dom'),
+    useNavigate: jest.fn()
+  };
+});
+
+describe('page/ProjectOverview/ProjectScore', () => {
+  const navigateSpy = jest.fn();
+  let request: jest.SpyInstance;
+
+  beforeEach(() => {
+    projectOverview.mockAllApi();
+    mockUseCurrentProject();
+    mockUseCurrentUser();
+    jest.useFakeTimers();
+    (useNavigate as jest.Mock).mockImplementation(() => navigateSpy);
+    request = projectOverview.getProjectScore();
+  });
+
+  afterEach(() => {
+    jest.useRealTimers();
+    jest.clearAllMocks();
+    cleanup();
+  });
+
+  const customRender = () => {
+    return superRender(<ProjectScore />);
+  };
+
+  it('render order status and click create order', async () => {
+    const { baseElement } = customRender();
+    await act(async () =>
+      EventEmitter.emit(EmitterKey.Refresh_Project_Overview)
+    );
+    await act(async () => jest.advanceTimersByTime(3000));
+    expect(request).toBeCalledWith({
+      project_name: mockProjectInfo.projectName
+    });
+    expect(baseElement).toMatchSnapshot();
+    expect(screen.getByText('项目评分')).toBeInTheDocument();
+    expect(screen.getByText('最新评分')).toBeInTheDocument();
+  });
+
+  it('render score between 60-80', async () => {
+    request.mockImplementation(() =>
+      createSpySuccessResponse({
+        data: {
+          score: 70
+        }
+      })
+    );
+    const { baseElement } = customRender();
+    await act(async () =>
+      EventEmitter.emit(EmitterKey.Refresh_Project_Overview)
+    );
+    await act(async () => jest.advanceTimersByTime(3000));
+    expect(request).toBeCalledWith({
+      project_name: mockProjectInfo.projectName
+    });
+    expect(baseElement).toMatchSnapshot();
+  });
+
+  it('render score between 40-60', async () => {
+    request.mockImplementation(() =>
+      createSpySuccessResponse({
+        data: {
+          score: 50
+        }
+      })
+    );
+    const { baseElement } = customRender();
+    await act(async () =>
+      EventEmitter.emit(EmitterKey.Refresh_Project_Overview)
+    );
+    await act(async () => jest.advanceTimersByTime(3000));
+    expect(request).toBeCalledWith({
+      project_name: mockProjectInfo.projectName
+    });
+    expect(baseElement).toMatchSnapshot();
+  });
+
+  it('render score less than 40', async () => {
+    request.mockImplementation(() =>
+      createSpySuccessResponse({
+        data: {
+          score: 30
+        }
+      })
+    );
+    const { baseElement } = customRender();
+    await act(async () =>
+      EventEmitter.emit(EmitterKey.Refresh_Project_Overview)
+    );
+    await act(async () => jest.advanceTimersByTime(3000));
+    expect(request).toBeCalledWith({
+      project_name: mockProjectInfo.projectName
+    });
+    expect(baseElement).toMatchSnapshot();
+  });
+
+  it('render error and refresh data', async () => {
+    const eventEmitSpy = jest.spyOn(eventEmitter, 'emit');
+    request.mockImplementation(() =>
+      createSpySuccessResponse({
+        code: 500,
+        message: 'error'
+      })
+    );
+    const { baseElement } = customRender();
+    await act(async () =>
+      EventEmitter.emit(EmitterKey.Refresh_Project_Overview)
+    );
+    await act(async () => jest.advanceTimersByTime(3000));
+    expect(request).toBeCalledWith({
+      project_name: mockProjectInfo.projectName
+    });
+    expect(baseElement).toMatchSnapshot();
+    expect(screen.getByText('error')).toBeInTheDocument();
+    expect(screen.getByText('刷新')).toBeInTheDocument();
+    fireEvent.click(screen.getByText('刷新'));
+    expect(eventEmitSpy).toBeCalledTimes(1);
+    expect(eventEmitSpy).toBeCalledWith(EmitterKey.Refresh_Project_Overview);
+  });
+});
